+++ kernel/util/bitmap.h
@@ -40,4 +40,4 @@ bool bitmap_test(const bitmap_t* bm, size_t index);
 void bitmap_set_all(bitmap_t* bm, size_t items);
 void bitmap_clear_all(bitmap_t* bm, size_t items);
 
-size_t bitmap_find_and_set(bitmap_t* bm, size_t items);
+long bitmap_find_and_set(bitmap_t* bm, size_t items);
